Minor bug fixes

+++ toolset/benchmark/benchmarker.py
@@ -540,7 +540,8 @@ class Benchmarker:
             logging.debug("Child process for {name} is still alive. Terminating.".format(name=test.name))
             self.__write_intermediate_results(test.name,"__run_test timeout (="+ str(self.run_test_timeout_seconds) + " seconds)")
             test_process.terminate()
-    os.remove('current_benchmark.txt')
+    if os.path.isfile('current_benchmark.txt'):
+      os.remove('current_benchmark.txt')
     logging.debug("End __run_tests.")
 
   ############################################################
